Refactor: Move KeyPairAndId to security module and rename it
