+++ desktop/src/main/java/bisq/desktop/main/content/user/user_profile/create/step2/CreateNewProfileStep2Model.java
@@ -17,9 +17,9 @@
 
 package bisq.desktop.main.content.user.user_profile.create.step2;
 
-import bisq.desktop.common.data.KeyPairAndId;
 import bisq.desktop.common.view.Model;
 import bisq.identity.Identity;
+import bisq.security.KeyIdKeyPairTuple;
 import bisq.security.pow.ProofOfWork;
 import javafx.beans.property.*;
 import javafx.scene.image.Image;
@@ -31,7 +31,7 @@ import java.util.Optional;
 @Getter
 public class CreateNewProfileStep2Model implements Model {
     @Setter
-    private Optional<KeyPairAndId> tempKeyPairAndId = Optional.empty();
+    private Optional<KeyIdKeyPairTuple> tempKeyPairAndId = Optional.empty();
     @Setter
     private Optional<ProofOfWork> proofOfWork = Optional.empty();
